# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/ricerca": {
      "post": {
        "operationId": "RicercaController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRicercaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ricerca created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseRicercaDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          }
        },
        "summary": "Create a new recruitment campaign",
        "tags": [
          "Ricerche"
        ]
      },
      "get": {
        "operationId": "RicercaController_findAll",
        "parameters": [
          {
            "name": "exclude",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "id",
                  "stato",
                  "descrizione",
                  "interviste",
                  "selezionati"
                ]
              }
            }
          },
          {
            "name": "fullRelations",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "number",
                "minimum": 0
              }
            }
          },
          {
            "name": "stato",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "interviste",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "number",
                "minimum": 0
              }
            }
          },
          {
            "name": "selezionati",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "number",
                "minimum": 0
              }
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of matching ricerche",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseRicercaDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Retrieve recruitment campaigns with filters and pagination",
        "tags": [
          "Ricerche"
        ]
      }
    },
    "/ricerca/{id}": {
      "get": {
        "operationId": "RicercaController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exclude",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "id",
                  "stato",
                  "descrizione",
                  "interviste",
                  "selezionati"
                ]
              }
            }
          },
          {
            "name": "fullRelations",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ricerca found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseRicercaDto"
                }
              }
            }
          },
          "404": {
            "description": "Ricerca not found"
          }
        },
        "summary": "Get a single recruitment campaign by ID",
        "tags": [
          "Ricerche"
        ]
      },
      "patch": {
        "operationId": "RicercaController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRicercaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ricerca updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseRicercaDto"
                }
              }
            }
          },
          "404": {
            "description": "Ricerca not found"
          }
        },
        "summary": "Update a recruitment campaign",
        "tags": [
          "Ricerche"
        ]
      },
      "delete": {
        "operationId": "RicercaController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ricerca deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseRicercaDto"
                }
              }
            }
          },
          "404": {
            "description": "Ricerca not found"
          }
        },
        "summary": "Delete a recruitment campaign by ID",
        "tags": [
          "Ricerche"
        ]
      }
    },
    "/ricerca/{id}/selezionato": {
      "post": {
        "operationId": "RicercaController_addSelezionato",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UtenteIdDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Candidate added to selection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseRicercaDto"
                }
              }
            }
          },
          "404": {
            "description": "Ricerca or user not found"
          }
        },
        "summary": "Add a selected candidate to this recruitment campaign",
        "tags": [
          "Ricerche"
        ]
      }
    },
    "/ricerca/{id}/selezionato/{utenteId}": {
      "delete": {
        "operationId": "RicercaController_removeSelezionato",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "utenteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Candidate removed from selection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseRicercaDto"
                }
              }
            }
          },
          "404": {
            "description": "Ricerca or user not found or not selected"
          }
        },
        "summary": "Remove a candidate from the selection",
        "tags": [
          "Ricerche"
        ]
      }
    }
  },
  "info": {
    "title": "IntervistaServer API",
    "description": "API specification for IntervistaServer: candidate management, recruitment searches, and conflict-validated interview scheduling.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UtenteIdDto": {
        "type": "object",
        "properties": {
          "utenteId": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "utenteId"
        ]
      },
      "CreateRicercaDto": {
        "type": "object",
        "properties": {
          "descrizione": {
            "type": "string"
          },
          "stato": {
            "type": "string"
          }
        },
        "required": [
          "descrizione"
        ]
      },
      "ResponseRicercaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "interviste": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "selezionati": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "descrizione": {
            "type": "string"
          },
          "stato": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "interviste",
          "selezionati",
          "descrizione"
        ]
      },
      "UpdateRicercaDto": {
        "type": "object",
        "properties": {
          "descrizione": {
            "type": "string"
          },
          "stato": {
            "type": "string"
          }
        }
      }
    }
  }
}